% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{integer. Length of the Fibonacci Sequence}
}
\value{
The Fibonacci Sequence
}
\description{
The Fibonacci Sequence is the series of numbers: 
0, 1, 1, 2, 3, 5, 8, 13, 21, 34, ... 
The next number is found by adding up the two numbers before it.
}
\details{
Implemented in C++
}
