% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Grade}
\usage{
grade(ids = NULL, file = NULL, verbose = TRUE)
}
\arguments{
\item{ids}{character. One or more repo ids to check. If \code{NULL} (the default) checks all repositories. See also \code{\link{repo}}}

\item{file}{character. Filename to store results (csv format)}

\item{verbose}{logical. Print on progress?}
}
\value{
data.frame
\describe{
  \item{id}{GitHub repo id}
  \item{pkg}{Name of the package}
  \item{grade}{PASS or FAIL}
}
}
\description{
Assigns a grade (pass/fail) to the projects for the exam "Coding for Data Science and Data Management", R module (2019/2020).
Master in Data Science and Economics, University of Milan.
}
\examples{
\dontrun{
 grade()
}

}
