% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repo}
\alias{repo}
\title{Ids of @unimi-dse GitHub repositories}
\format{character vector of ids}
\source{
\url{https://github.com/unimi-dse}
}
\usage{
repo
}
\description{
A dataset containing the ids for the "Coding for Data Science and Data Management" exam, R module (2019/2020).
Master in Data Science and Economics, University of Milan.
}
\keyword{datasets}
