% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sampleGUI}
\alias{sampleGUI}
\title{Sample GUI}
\usage{
sampleGUI()
}
\value{
shiny app
}
\description{
Runs a Graphical User Interface to evaluate the projects for the exam "Coding for Data Science and Data Management", R module (2019/2020).
Master in Data Science and Economics, University of Milan.
}
